[P2PW-1134] - Additinal Moonpay fixes for feature builds (#196)
* [P2PW-1134] - Additinal Moonpay fixes for feature builds

* [P2PW-1134] - Final Moonpay fixes for feature builds

* [P2PW-1134] - Amount fix for moonpay

+++ app/src/main/java/org/p2p/wallet/main/api/MoonpayBuyCurrencyResponse.kt
@@ -13,7 +13,7 @@ data class MoonpayBuyCurrencyResponse(
     @SerializedName("networkFeeAmount")
     val networkFeeAmount: BigDecimal,
     @SerializedName("quoteCurrencyAmount")
-    val quoteCurrencyAmount: BigDecimal,
+    val quoteCurrencyAmount: Double,
     @SerializedName("totalAmount")
     val totalAmount: BigDecimal
 )
